Improve error handling in service classes
Create an Unprocessable{Redditor,Thread} instance
when the reddit or openai API raises an exception.
The Redditor or Thread that was being processed when
the error was thrown will attempt to be reprocessed
after the amount of time as specified by
UNPROCESSABLE_REDDITOR_EXP_TD or UNPROCESSABLE_THREAD_EXP_TD
has elapsed and the username or url gets resubmitted.