Add listeners to auto set current context filter
Add listener to automatically detect when the page/tab URL
changes and then get the current context from that URL.
Then lookup if there is a defined content filter that matches
the found context and store that filter in storage.
Anywhere that needs to reference the current context
filter can just lookup the ACTIVE_CONTENT_FILTER
storage key and it will be accurate.