Modify routing to fix setting nav tabs
React router defaults to "/" when the popup opens.
NavLinks on the Settings view are mapped to /settings
and /status. That means that the active tab
displayed for the Settings view would not show
correctly when the popup opened if the user was
authenticated. It would show correctly after login
because they would be redirected to /settings, but
if the popup was opened when authentication credentials
were already present the current route would be "/".