Fix bug for status messages where status is computed
If the status message active field is computed, the
active attribute must be hardcoded to inverse of
whatever the initial value of the constance config
key is set to. This is because a signal will not
be dispatched by constance when the app starts up.
A signal will only be dispatched when an admin
changes the constance config in the admin panel
manually.