Set extension variables based on API status messages
Return all messages from status messages endpoint
instead of just active messages. When showing
messages in the extension, filter to show the
ones that are active. Use the status message active
attribute to set variables stored by the extension.
This allows us to modify the behavior of the
extension based on changes to admin settings in
the API.