Fix bug causes duplicate old data spans to stay rendered
If a pending span is present in the DOM and then scrolling
causes the processed result to get fetched, the processed
result was being rendered in addition to the span that
said the results were pending. Delete the old span for
each thread of redditor before annotating with results
from the latest request.