Do not store context-query job ID in extension storage
If the use navigates away from the context query form
while the query is still being processed, the result
will still be accessible in the query history. The
original intent for storing it in extension storage
was to ensure that the result could still be fetched
if the user refreshed the page.