Do away with prompt prefix constance setting
It made sense for context queries when the user's
query would be appended to the prefix. However,
user's will soon see a prepopulated default context
query prompt that will include the old prefix. They
will be able to modify the prepopulated string however
they wish and then that is what will be passed to
the LLM when the context query is processed.