Rework how llm prompts are passed around
I did not like how I did it before as it seems prone
to error in the long run. Simply store the distinct
prompts and access them on those prompts. When processing
context-queries via management commands, provide the
prompt as an editable default value in the prompt
to mimic the behavior that will be available to users when
submitting context-queries via the extension.