Fix context-query producer bug
If a user submitted a context query for an entity
whose data has not been processed yet, data processing
will occur before the context query is processed.
However, if the user's LLM selection was being used
to process the entity's data. This should not happen
because all entities should be processed according
to the LLM defined in the admin config. Needed to
add a separate variable to differentiate between
the LLM producer that should be used for context-
query processing vs data processing.