Fix bug for default content filter
If no custom content filter is found, the default
filter is used. However, the bug was that the default
content filter was being referenced by the constant
value which is wrong because the default content
filter can be modified by the user. The default
content filter in constants does not change.