Do not allow OpenAI downtime to cause reecon downtime
If the OpenAI API is throwing errors, do not allow their
downtime to cause downtime in reecon. Assume the provided
api key is valid so that the reecon API can still serve
processed data.

We can assume the api key is valid because when the
user initially enters it in the extension, it will
not be stored unless it passes a validation test.