Add thread filter table
Make existing filter table more generic so it can
be used for both the comment and thread filter settings.

Set current context to separate variables for comments
and threads so that a matching filter defined for one and
not the other can be active while the other uses the default
filter.