Change redditor account min age from 7 days to 1 hour
I am not sure if this is a feature I want to keep
so I am drastically reducing the account min age
for now. Sometimes a thread will be posted from a
throwaway account and it would still be interesting
to process that user. They would still end up being
processed if the same thread was viewed in the
future anyways.