Update dependencies
There is currently a bug that prevents the extension
from building without error. I created a new plasmo
extension, installed all the same dependencies, and
updated the versions in this repo to the versions
used in the fresh plasmo extension. The build problem
is still present, but at least all the dependencies
are consistent now.