Improve filtering of reddit submissions
After performing text transformations that require
markdown to be present (like removing block quotes),
strip all markdown from the text so it can be more
predictably processed. For example, a string containing an
embedded URL should be transformed to only be the
string seen by the user. We can can then test if
that string meets our requirements such as min and
max length.