Bump extension dependencies
MUI is pinned to 5.16.14 as 6.0.0 and beyond cause an
error to be thrown when using the extension that breaks
it - https://github.com/mui/material-ui/issues/43477

Also add back jwt-decode
- I removed it as an external
dependencies but forgot to re-add it to package.json
after undoing that change.

Also replace marked with mui-markdown package
- The new package simplifies the usage of markdown
by not having to also create a custom component.