Re-work mechanism to show/hide filtered content
I tried several approaches to hide filtered content
from being visible including:
- Overlaying an element on the content to hide
  - Cannot use plasmo overlay because it does not
    react to dynamically loaded content (infinite
    scroll)
- Make filtered comments appear the same as "spoiler
  text" that becomes visible onclick
  - Does not seem like a good approach to use
    pre-existing behavior from the source page for
    purposes other than they were intended for.
- Collapse filtered comments
  - Also collapses all child comments which likely
    would not have been filtered if they were top
    level comments.

The approach I settles on is to simply set the div
containing the comment text to `style.display = 'none'`
for filtered comments and 'block' for comments which
should be visible. If the user wants to see a filtered
comment, they can click on the data icon to toggle it's
visibility.