Rework models, use langchain for LLM queries
Originally, each column of generated data could have
been generated by an LLM or a NLP source. This is why
all data had a foreign key to the source that generated
it. This is no longer the case as everything is now
generated by the same LLM source for the given job.
The models were simplified to not duplicate the source
information for each piece of generated data. This
reduces the complexity of queries and amount of data
stored.

Replace openai package with langchain as it provides
a common interface to multiple LLM providers. It also
provides a convenient way to retrieve metadata about
each query regarding the number of input and output
tokens used for the query.